add tests for coordinate conversion functions
